requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
pandas
numpy
scipy
h5py
matplotlib
colorama
tqdm
pyarrow
terminaltables
py-cpuinfo==8.0.0
scikit-learn-intelex
# imbalanced-learn
Pillow
pyparsing
omegaconf
hydra-core==1.2.0
torch==1.13.0
# torchaudio==0.14.0
# torchaudio==0.13.0 # 0.14.0 is not available for some reason for me
torchmetrics==0.10.3
torchvision==0.14.0
# PyYAML==6.0
pytorch-lightning==1.8.3.post1
opencv-contrib-python
scikit-image
cvxopt
kornia